Added "Insert new object" button in SceneCanvas when right clicking. Reworked compilation dialog. Adapted to changes in GDCore.
git-svn-id: svn://localhost@816 8062f311-0dae-4547-b526-b8ab9ac864a5

+++ IDE/SceneCanvasEditionData.h
@@ -43,8 +43,8 @@ public:
     float yEndRectangleSelection;
 
     //Remember cursor old position
-    int oldMouseX;
-    int oldMouseY;
+    float oldMouseX;
+    float oldMouseY;
 
     //Déplacement avec la souris ( clic molette )
     bool isMoving;
